Move diff functions to separate file: utils/results/diff.go
