+++ utils/results/results.go
@@ -3,8 +3,6 @@ package results
 import (
 	"errors"
 	"fmt"
-	"path/filepath"
-	"strings"
 	"sync"
 	"time"
 
@@ -732,352 +730,4 @@ func (jsr *JasScansResults) HasInformationByType(scanType jasutils.JasScanType) 
 		}
 	}
 	return false
-}
-
-// UnifyScaAndJasResults merges SCA and JAS diff results into a single SecurityCommandResults.
-func UnifyScaAndJasResults(scaResults, jasDiffResults *SecurityCommandResults) *SecurityCommandResults {
-	// Create unified results based on JAS diff structure
-	unifiedResults := &SecurityCommandResults{
-		ResultsMetaData: ResultsMetaData{
-			EntitledForJas:   jasDiffResults.EntitledForJas,
-			SecretValidation: jasDiffResults.SecretValidation,
-			CmdType:          jasDiffResults.CmdType,
-			XrayVersion:      jasDiffResults.XrayVersion,
-			XscVersion:       jasDiffResults.XscVersion,
-			MultiScanId:      jasDiffResults.MultiScanId,
-			StartTime:        jasDiffResults.StartTime,
-			ResultContext:    jasDiffResults.ResultContext,
-		},
-	}
-
-	// Merge targets from both SCA and JAS results
-	for _, scaTarget := range scaResults.Targets {
-		// Find corresponding JAS target
-		var jasTarget *TargetResults
-		for _, jTarget := range jasDiffResults.Targets {
-			if jTarget.Target == scaTarget.Target {
-				jasTarget = jTarget
-				break
-			}
-		}
-
-		// Create unified target with both SCA and JAS results
-		unifiedTarget := &TargetResults{
-			ScanTarget:       scaTarget.ScanTarget,
-			AppsConfigModule: scaTarget.AppsConfigModule,
-			ScaResults:       scaTarget.ScaResults,
-			JasResults:       nil,
-		}
-
-		// Add JAS diff results if available
-		if jasTarget != nil {
-			unifiedTarget.JasResults = jasTarget.JasResults
-		}
-
-		unifiedResults.Targets = append(unifiedResults.Targets, unifiedTarget)
-	}
-
-	return unifiedResults
-}
-
-// CompareJasResults computes the diff between target and source JAS results.
-// Returns only NEW findings in source that don't exist in target.
-func CompareJasResults(targetResults, sourceResults *SecurityCommandResults) *SecurityCommandResults {
-	log.Info("[DIFF] Starting JAS diff calculation")
-	log.Debug("[DIFF] Comparing", len(sourceResults.Targets), "source targets against", len(targetResults.Targets), "target targets")
-
-	// Create diff results based on source structure
-	diffResults := &SecurityCommandResults{
-		ResultsMetaData: ResultsMetaData{
-			EntitledForJas:   sourceResults.EntitledForJas,
-			SecretValidation: sourceResults.SecretValidation,
-			CmdType:          sourceResults.CmdType,
-			XrayVersion:      sourceResults.XrayVersion,
-			XscVersion:       sourceResults.XscVersion,
-			MultiScanId:      sourceResults.MultiScanId,
-			StartTime:        sourceResults.StartTime,
-			ResultContext:    sourceResults.ResultContext,
-		},
-	}
-
-	// Compare each source target against ALL target targets
-	for _, sourceTarget := range sourceResults.Targets {
-		if sourceTarget.JasResults == nil {
-			continue
-		}
-
-		// Collect ALL target JAS results to compare against
-		var allTargetJasResults []*JasScansResults
-		for _, targetTarget := range targetResults.Targets {
-			if targetTarget.JasResults != nil {
-				allTargetJasResults = append(allTargetJasResults, targetTarget.JasResults)
-			}
-		}
-
-		diffJasResults := filterExistingFindings(allTargetJasResults, sourceTarget.JasResults)
-
-		diffTarget := &TargetResults{
-			ScanTarget: sourceTarget.ScanTarget,
-			JasResults: diffJasResults,
-		}
-
-		diffResults.Targets = append(diffResults.Targets, diffTarget)
-	}
-
-	return diffResults
-}
-
-// filterExistingFindings removes findings from source that already exist in target.
-func filterExistingFindings(allTargetJasResults []*JasScansResults, sourceJasResults *JasScansResults) *JasScansResults {
-	if sourceJasResults == nil {
-		return nil
-	}
-
-	// If no target results, return all source results (everything is new)
-	if len(allTargetJasResults) == 0 {
-		return sourceJasResults
-	}
-
-	// Build target fingerprint set from ALL target results
-	targetKeys := make(map[string]bool)
-
-	for _, targetJasResults := range allTargetJasResults {
-		if targetJasResults == nil {
-			continue
-		}
-
-		// Extract from target secrets results (location-based)
-		for _, targetRun := range targetJasResults.GetVulnerabilitiesResults(jasutils.Secrets) {
-			extractLocationsOnly(targetRun, targetKeys)
-		}
-		for _, targetRun := range targetJasResults.GetViolationsResults(jasutils.Secrets) {
-			extractLocationsOnly(targetRun, targetKeys)
-		}
-		// Extract from target IaC results (location-based)
-		for _, targetRun := range targetJasResults.GetVulnerabilitiesResults(jasutils.IaC) {
-			extractLocationsOnly(targetRun, targetKeys)
-		}
-		for _, targetRun := range targetJasResults.GetViolationsResults(jasutils.IaC) {
-			extractLocationsOnly(targetRun, targetKeys)
-		}
-		// Extract from target SAST results (fingerprint-based)
-		for _, targetRun := range targetJasResults.GetVulnerabilitiesResults(jasutils.Sast) {
-			extractFingerprints(targetRun, targetKeys)
-		}
-		for _, targetRun := range targetJasResults.GetViolationsResults(jasutils.Sast) {
-			extractFingerprints(targetRun, targetKeys)
-		}
-	}
-
-	log.Debug("[DIFF] Built target fingerprint set with", len(targetKeys), "unique keys")
-
-	// Count source results before filtering
-	sourceSecrets := countSarifResults(sourceJasResults.JasVulnerabilities.SecretsScanResults) +
-		countSarifResults(sourceJasResults.JasViolations.SecretsScanResults)
-	sourceIac := countSarifResults(sourceJasResults.JasVulnerabilities.IacScanResults) +
-		countSarifResults(sourceJasResults.JasViolations.IacScanResults)
-	sourceSast := countSarifResults(sourceJasResults.JasVulnerabilities.SastScanResults) +
-		countSarifResults(sourceJasResults.JasViolations.SastScanResults)
-
-	log.Debug("[DIFF] Source findings before diff - Secrets:", sourceSecrets, "| IaC:", sourceIac, "| SAST:", sourceSast)
-
-	// Filter source results - keep only what's NOT in target
-	filteredJasResults := &JasScansResults{}
-
-	// Filter vulnerabilities
-	filteredJasResults.JasVulnerabilities.SecretsScanResults = filterSarifRuns(
-		sourceJasResults.JasVulnerabilities.SecretsScanResults, targetKeys)
-	filteredJasResults.JasVulnerabilities.IacScanResults = filterSarifRuns(
-		sourceJasResults.JasVulnerabilities.IacScanResults, targetKeys)
-	filteredJasResults.JasVulnerabilities.SastScanResults = filterSarifRuns(
-		sourceJasResults.JasVulnerabilities.SastScanResults, targetKeys)
-
-	// Filter violations
-	filteredJasResults.JasViolations.SecretsScanResults = filterSarifRuns(
-		sourceJasResults.JasViolations.SecretsScanResults, targetKeys)
-	filteredJasResults.JasViolations.IacScanResults = filterSarifRuns(
-		sourceJasResults.JasViolations.IacScanResults, targetKeys)
-	filteredJasResults.JasViolations.SastScanResults = filterSarifRuns(
-		sourceJasResults.JasViolations.SastScanResults, targetKeys)
-
-	// Count filtered results after diff
-	diffSecrets := countSarifResults(filteredJasResults.JasVulnerabilities.SecretsScanResults) +
-		countSarifResults(filteredJasResults.JasViolations.SecretsScanResults)
-	diffIac := countSarifResults(filteredJasResults.JasVulnerabilities.IacScanResults) +
-		countSarifResults(filteredJasResults.JasViolations.IacScanResults)
-	diffSast := countSarifResults(filteredJasResults.JasVulnerabilities.SastScanResults) +
-		countSarifResults(filteredJasResults.JasViolations.SastScanResults)
-
-	log.Info("[DIFF] New findings after diff - Secrets:", diffSecrets, "| IaC:", diffIac, "| SAST:", diffSast)
-	log.Info("[DIFF] Filtered out - Secrets:", sourceSecrets-diffSecrets, "| IaC:", sourceIac-diffIac, "| SAST:", sourceSast-diffSast)
-
-	return filteredJasResults
-}
-
-// countSarifResults counts total results across all SARIF runs
-func countSarifResults(runs []*sarif.Run) int {
-	count := 0
-	for _, run := range runs {
-		if run != nil {
-			count += len(run.Results)
-		}
-	}
-	return count
-}
-
-// extractFingerprints extracts fingerprints from SARIF run (for SAST)
-func extractFingerprints(run *sarif.Run, targetKeys map[string]bool) {
-	for _, result := range run.Results {
-		if result.Fingerprints != nil {
-			key := getResultFingerprint(result)
-			if key != "" {
-				targetKeys[key] = true
-			}
-		} else {
-			for _, location := range result.Locations {
-				key := getRelativeLocationFileName(location, run.Invocations) + getLocationSnippetText(location)
-				targetKeys[key] = true
-			}
-		}
-	}
-}
-
-// extractLocationsOnly extracts locations (for Secrets and IaC - no fingerprints)
-func extractLocationsOnly(run *sarif.Run, targetKeys map[string]bool) {
-	for _, result := range run.Results {
-		for _, location := range result.Locations {
-			key := getRelativeLocationFileName(location, run.Invocations) + getLocationSnippetText(location)
-			targetKeys[key] = true
-		}
-	}
-}
-
-// getResultFingerprint returns the SAST fingerprint from a result
-func getResultFingerprint(result *sarif.Result) string {
-	if result.Fingerprints != nil {
-		if value, ok := result.Fingerprints["precise_sink_and_sink_function"]; ok {
-			return value
-		}
-	}
-	return ""
-}
-
-// getLocationSnippetText returns the snippet text from a location
-func getLocationSnippetText(location *sarif.Location) string {
-	if location.PhysicalLocation != nil && location.PhysicalLocation.Region != nil &&
-		location.PhysicalLocation.Region.Snippet != nil && location.PhysicalLocation.Region.Snippet.Text != nil {
-		return *location.PhysicalLocation.Region.Snippet.Text
-	}
-	return ""
-}
-
-// getRelativeLocationFileName returns the relative file path from a location
-func getRelativeLocationFileName(location *sarif.Location, invocations []*sarif.Invocation) string {
-	wd := ""
-	if len(invocations) > 0 {
-		wd = getInvocationWorkingDirectory(invocations[0])
-	}
-	filePath := getLocationFileName(location)
-	if filePath != "" {
-		return extractRelativePath(filePath, wd)
-	}
-	return ""
-}
-
-func getInvocationWorkingDirectory(invocation *sarif.Invocation) string {
-	if invocation != nil && invocation.WorkingDirectory != nil && invocation.WorkingDirectory.URI != nil {
-		return *invocation.WorkingDirectory.URI
-	}
-	return ""
-}
-
-func getLocationFileName(location *sarif.Location) string {
-	if location != nil && location.PhysicalLocation != nil && location.PhysicalLocation.ArtifactLocation != nil && location.PhysicalLocation.ArtifactLocation.URI != nil {
-		return *location.PhysicalLocation.ArtifactLocation.URI
-	}
-	return ""
-}
-
-func extractRelativePath(resultPath string, projectRoot string) string {
-	// Remove OS-specific file prefix
-	resultPath = strings.TrimPrefix(resultPath, "file:///private")
-	resultPath = strings.TrimPrefix(resultPath, "file:///")
-	projectRoot = strings.TrimPrefix(projectRoot, "file:///private")
-	projectRoot = strings.TrimPrefix(projectRoot, "file:///")
-	projectRoot = strings.TrimPrefix(projectRoot, "/")
-
-	// Get relative path (removes temp directory)
-	relativePath := strings.ReplaceAll(resultPath, projectRoot, "")
-	trimSlash := strings.TrimPrefix(relativePath, string(filepath.Separator))
-	return strings.TrimPrefix(trimSlash, "/")
-}
-
-// MergeStatusCodes merges two ResultsStatus structs, taking the worst (non-zero) status for each scanner
-// This is used when combining target and source results to ensure partial results filtering works correctly
-func MergeStatusCodes(target, source ResultsStatus) ResultsStatus {
-	merged := ResultsStatus{}
-	merged.SbomScanStatusCode = mergeStatusCode(target.SbomScanStatusCode, source.SbomScanStatusCode)
-	merged.ScaScanStatusCode = mergeStatusCode(target.ScaScanStatusCode, source.ScaScanStatusCode)
-	merged.ContextualAnalysisStatusCode = mergeStatusCode(target.ContextualAnalysisStatusCode, source.ContextualAnalysisStatusCode)
-	merged.SecretsScanStatusCode = mergeStatusCode(target.SecretsScanStatusCode, source.SecretsScanStatusCode)
-	merged.IacScanStatusCode = mergeStatusCode(target.IacScanStatusCode, source.IacScanStatusCode)
-	merged.SastScanStatusCode = mergeStatusCode(target.SastScanStatusCode, source.SastScanStatusCode)
-	merged.ViolationsStatusCode = mergeStatusCode(target.ViolationsStatusCode, source.ViolationsStatusCode)
-	return merged
-}
-
-// mergeStatusCode returns the worst (non-zero) status code between two
-func mergeStatusCode(a, b *int) *int {
-	if a == nil {
-		return b
-	}
-	if b == nil {
-		return a
-	}
-	// Return the non-zero value (failed status), or zero if both succeeded
-	if *a != 0 {
-		return a
-	}
-	return b
-}
-
-// filterSarifRuns filters SARIF runs, keeping only results that are NOT in target
-func filterSarifRuns(sourceRuns []*sarif.Run, targetKeys map[string]bool) []*sarif.Run {
-	var filteredRuns []*sarif.Run
-
-	for _, run := range sourceRuns {
-		var filteredResults []*sarif.Result
-
-		for _, result := range run.Results {
-			if result.Fingerprints != nil {
-				// Use fingerprint for matching (SAST)
-				if !targetKeys[getResultFingerprint(result)] {
-					filteredResults = append(filteredResults, result)
-				}
-			} else {
-				// Use location for matching (Secrets, IaC)
-				var filteredLocations []*sarif.Location
-				for _, location := range result.Locations {
-					key := getRelativeLocationFileName(location, run.Invocations) + getLocationSnippetText(location)
-					if !targetKeys[key] {
-						filteredLocations = append(filteredLocations, location)
-					}
-				}
-
-				if len(filteredLocations) > 0 {
-					newResult := *result
-					newResult.Locations = filteredLocations
-					filteredResults = append(filteredResults, &newResult)
-				}
-			}
-		}
-
-		if len(filteredResults) > 0 {
-			filteredRun := *run
-			filteredRun.Results = filteredResults
-			filteredRuns = append(filteredRuns, &filteredRun)
-		}
-	}
-
-	return filteredRuns
 }
